+++ src/main/java/lotto/LottoCustomer.java
@@ -1,9 +1,11 @@
 package lotto;
 
+import lotto.domain.LottoSeller;
+import lotto.domain.Lottos;
 import lotto.view.InputView;
 import lotto.view.OutputView;
 
-public class LottoUser {
+public class LottoCustomer {
 
     public static void main(String[] args){
         int price = InputView.requestPrice();
