+++ src/test/java/lotto/LottoTest.java
@@ -8,6 +8,11 @@ import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 import java.util.stream.Stream;
+import lotto.domain.Lotto;
+import lotto.domain.LottoNum;
+import lotto.domain.LottoNums;
+import lotto.domain.LottoSeller;
+import lotto.domain.Lottos;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
